backend/GetNextQuestion/requirements.txt
requests
boto3
firebase-admin
google-cloud-pubsub